requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
wikipedia>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
streamlit>=1.28.0
faiss-cpu>=1.7.4 